[feat] Migrate order details summary to independent component
