+++ src/components/SummaryCart/SummaryCart.module.scss
@@ -1,0 +1,41 @@
+.root {
+  .summary {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background-color: var(--bg-tertiary);
+    padding: 2em 1em 2em 1em;
+    position: relative;
+    z-index: 6;
+
+    p {
+      color: var(--text-tertiary);
+      font-size: 1.25rem;
+
+      img {
+        transform: rotate(0deg);
+        transition: 0.6s;
+        &.active {
+          transform: rotate(180deg);
+        }
+      }
+    }
+
+    .amount {
+      font-weight: 500;
+      font-size: 1.25rem;
+    }
+  }
+
+  .products {
+    transform: translateY(-200%);
+    position: absolute;
+    background-color: var(--bg-primary);
+    transition: transform 0.6s ease-in-out;
+    z-index: 5;
+
+    &.active {
+      transform: translateY(0);
+    }
+  }
+}
