+++ src/pages/Shipping/index.tsx
@@ -1,7 +1,15 @@
-import React from "react";
+import { useSelector } from "react-redux";
+import Breadcrumbs from "../../components/Breadcrumbs/Breadcrumbs";
+import OrderDetailsProduct from "../../components/SummaryCart/OrderDetailsProduct/OrderDetailsProduct";
 
 const index = () => {
-  return <div>Shipping</div>;
+  const data = useSelector((state: any) => state);
+
+  return (
+    <main>
+      <Breadcrumbs />
+    </main>
+  );
 };
 
 export default index;
